src/text_to_image_service/requirements.txt
flask 
pyjwt
pylint
jedi
flask_sqlalchemy
astroid
click
isort
itsdangerous
Jinja2
lazy-object-proxy
MarkupSafe
mccabe
SQLAlchemy 
marshmallow
psycopg2
parso
platformdirs
toml
werkzeug
wrapt
requests
openai
environ
django-environ==0.11.2
alembic==1.13.1
PyMongo==4.7.2
flask_pymongo==2.3.0
python-dotenv==1.0.1
nameko==2.14.1